Women traumatized by violence
This work developed during the last 35 (thirty-five years) began as an intervention in the field in 1983 in favelas of Rio de Janeiro. At first, the calls for interventions were addressed to IBRAPSI (Brazilian Institute of Psychoanalysis, Groups and Institutions) by Resident Associations.
These interventions were carried out in the following places: Favela da Mangueira in the Maracanã Neighborhood, Favela do Jacarezinho, in the Bairro da Maria da Graça, Favela Chapéu da Mangueira, in the Leme Neighborhood.
These interventions began with families requesting help understanding their child's symptoms.
Two years later it became an intervention with women who suffered domestic violence. This change has come about because of the growing absence of parents, and because our approach is centered on the understanding that the symptoms presented by children arise and stratify themselves in the relationship with the parents first and secondly by influence with the environment they live.
From that point on we began to treat our work as action research, Lewin (1969). In 1987 we call this research participatory research (Saidon, 1983, Barbier, 1985, Thiollent, 1987 and Brandão, 1987). This form of research is also an action that aims to propose changes in the reality of the participants in effective social collaboration, giving the protagonism to the members of the research who were not researchers.
A proposal was made for exploratory methodologies, with objectives defined in the field of action by researchers and participants. Their results were linked to the awareness of the factors involved in immediate life situations and collective participation through the support of each member of the group.
The interferences, after the identifications of problems, were based on some classic exercises of group dynamics and the method of Augusto Boal. The data were treated as provisional as it applies to this type of research and the small and medium size groups, ranging from eight to ten women in each group in a total of 5 groups until 1987.
From that year the coordination left the Institution and happened to be of two researchers who continued the work. With the increase of violence in these localities the researchers were prohibited from continuing their work in the headquarters of Jacarezinho and Chapéu Mangueira Residents' Associations. In Mangueira we had the support of two influential residents.
The work with Mangueira only required the direct intervention of one researcher and the other was in charge of the analysis of the data.
In 1989, one researcher left health problems with only one. At that moment the women of Mangueira preferred to meet in another place. The chosen place was a dependency of UERJ (University of the State of Rio de Janeiro).
This work has undergone several changes both in participants and in technical approaches.
Now part of the concept of social cartography, where we seek in the territories and in their multiple interrelationships the possible relations with the institutions that are constantly changing in our country. We seek forms of institutional support and mainly forms of support outside institutions.
We are based on the concept of health of 2007, Vélez that integrates the individual welfare conditions, the right to quality of life and access to basic services that guarantee the development of these fundamental freedoms. The strengthening and enjoyment of these freedoms provides the individual with a set of skills to try to do and to be, and to become in life according to their ideals, without measuring obstacles generated by economic conditions, social, cultural situations and, in short, situations that are prevent their full realization and enjoyment of using their skills.
Without even the existence of public policies as governmental actions aimed at solving public and collective needs there is a vacuum in terms of assistance to victims of violence. In general, the state is being called to take responsibility in this world scenario where social responsibility is equally divided between it and society.
In this scenario of extreme precariousness we enter with our work to at least help in reducing the effects produced by the traumas caused by violence.
As we work in groups and leave tools for groups and women to go it alone, we close our participation with the establishment of resilient and cohesive emotional and affective forces.
Groups becoming cohesive are able to neutralize psychic waste caused by violence and inequalities.
In this work, we add some of the therapeutic principles of W. Reich, Shapiro and the theoretical foundations of Socioanalysis to the psychotherapies of body, we also align the predictors of resilience (facing adversities rather than re-creating suffering).
This therapy of better emotional and cognitive restructuring We used heart rate and skin impedance as a measure. Their results are stable.
We also seek to report on data on violence against women in Brazil:
 Brazil is ranked 5th among the 84 countries that have data on these crimes, after El Salvador, Columbia, Guatemala and the Russian Federation.
 In 2003: 3937 murders of women
 In 2013: 4762 murders of women
 The number of homicides of white women falls from 1747 victims in 2003 to 1576 in 2013
 The number of homicides of black women increases from 1864 to 2875 victims (or 54.2%).

Júlio Jacobo MAP OF VIOLENCE 2015

 

 
 


Brandão, F., Lyra da Silva, J. P. Reich the focal therapy precursor
Research: Socioeconomic Conditions and Domestic and Family Violence against Women
http://www.onumulheres.org.br/wp-content/uploads/2017/11/